fix: cast trash_retention_days to int for Carbon::addDays()
config() returns string from env; Carbon::rawAddUnit() requires int|float.
TypeError: Argument #3 must be of type int|float, string given.

+++ app/Http/Controllers/MediaController.php
@@ -97,7 +97,7 @@ class MediaController extends Controller
 
         $media->update([
             'trashed_at'  => now(),
-            'purge_after' => now()->addDays(config('gallery.trash_retention_days', 30)),
+            'purge_after' => now()->addDays((int) config('gallery.trash_retention_days', 30)),
         ]);
 
         AuditLog::record('media.trash', $media);
